Added a facility to add ocaml libraries to coccinelle.
Add your library to the list on the top of the configure file, which
specifies also if the library should be included by default. Otherwise,
you can use --with-$lib and --without-$lib to toggle its use.
Also, specify a configuration for your library in Makefile.libs, which
lists a number of variables in which you can specify which modules to
use, etc. If you want to optionally
use a local version of a library, make sure that you put it in a
subdirectory with exactly the same name as the library.

Currently, this mechanism runs alongside the existing library parameters.

svn path=/coccinelle/; revision=13244