Parallel build explicitly disabled for the code generation.
The ML-dependencies are not properly enforced in the Makefile. Running make in parallel can lead to the problem that
some interface files are recompiled (especially when you build both the opt and non-opt version), which can lead to
compilation failures.

We should switch to a tool like "ocamlbuild" in order to have proper dependency management and parallel builds.