removes the -thread and -dtypes command when dynamic linking ocaml scripts.
The -dtypes option leaves inferred type information of the generated file in
a .annot file, which seems not useful as this file should not be edited
directly. This option might still be useful if the information would be
in terms of the .cocci file and there would be support for it in an editor.