Adds an ocamlbuild configuration for building spatch.
It is not yet integrated into the Makefiles. It'll only
need a little Makefile support for unpacking and patching
the appropriate bundled packages. So, it cannot be released
yet.

It can be tested, however, if you first do a normal build
to prepare the sources:

> ./configure
> make && make clean

Now you can e.g. run:

> ocamlbuild main.byte
> ocamlbuild main.native

Which compiles the byte and native version of spatch, which
can subsequently be found in the _build directory.