Coccinelle release 1.0.0-rc19
A new version of Coccinelle is available from the web page, and probably
soon from github.  Some new features are as follows:

* Improved indentation and newline introduction in pretty printer.

* Elimination of the reliance on grep for selecting files to process.
This makes it substantially faster to run Coccinelle on code bases for
which no indexing has been done, and gives some performance improvement
when indexing is used as well.

* Allow pattern matching on the contents of strings.  An example is found
in demos/format.cocci.  There are two new kinds of metavariables: format
and format list.  Format matches a single format description after a %.
Format list matches a sequence of format declarations (including the %)
and the intervening string text.  It is possible to specify the length of
a format list, just as one can specify the length of an expression list.