Bug #5: Support for file display with special characters in the name
The command "cat" was called to display a file content in a debug mode.
The corresponding file name was passed in a way that the used parameter could
be split into separate parts if it contained special characters like spaces.

A name split could result in unwanted consequences.
1. Different files would be referenced.
2. The files would not be found with shorter names.
3. Trailing text could eventually be interpreted as further shell commands
   with unknown side effects.

This situation was fixed in the way that the specified file name will always
be handled as a single parameter for the desired task. The content display will
be performed by OCaml functions instead of an external program.

Signed-off-by: Markus Elfring <elfring@users.sourceforge.net>