Bug #5: Restrict a file name to naming convention for OCaml modules
External tools will be called for the transformation of an OCaml source file
into an executable program. The corresponding file name was treated in a way
that the used parameter could be split into separate parts if it contained
special characters like spaces.

A name split could result in unwanted consequences.
- Different files would be referenced.
- The files would not be found with shorter names.

This situation was improved in the way that a specific identifier which will
fit to the naming convention for OCaml modules will be generated.
Such a replacement should be safer now.

Signed-off-by: Markus Elfring <elfring@users.sourceforge.net>