This fixes two problems:
1. optim1 for function bodies that are just ... was ignoring the possible
when forall or when exists annotation on the dots.

2. dots_and_nests, as well as optim1, which is a simplified version of
dots_and_nests, was allowing After to be part of an exists path, which
could end up avoiding some when != annotated things, due to the after
that skips from the beginning to the end of a block.

The function get_quantifier is also created, to allow that code to be
called from both dots_and_nests and from the optim1 code.