This makes two major changes that are intertwined
1. Introduction of escape field of the TrueNode CFG node.  The purpose of
this field is to know whether a true branch only escapes, making it possible
error handling code.  This is determined in control_flow_c_build and used
in asttoctl2.ml, in the treatment of dots.  This escaping information
makes the detection of error exits much more efficient, because the CTL
for that only needs to be started if the branch aborts.

At the same time, a field was added to after to indicate whether the escape
is via a return, goto, etc.  This field is not used, but it is kept in case
it turns out to be useful in the future.

2. mkafter function in control_flow_c_build.  This function is used to add
after nodes.  Such nodes are added early in exits such as return, goto, etc.
If there is a normal exit from the compound or if, however, the early ones
are dropped.  The mkafter function thus makes the nodes and collects them
in a list so that they can be dropped later if not needed.  The main
purpose of all this is to be able to add things eg after a block that aborts.
But the same information turns out to also be useful in detecting error
exits.

Break nodes are also extended to distinguish between switch breaks and
loop breaks.  Switch breaks should not correspond to error handling code, even
if at the end of an if.