Take into account rule dependencies when considering plus information in get_constants.  Also removing constraints based on plus information results in True, regardless on whether & or v is being considered.
This requires adding Not to the formulas in get_constants2, to allow
making formulas that precisely reflect the provided dependencies for
a given rule.  This is a bit unfortunate because the formulas are used
for two things: these rule dependencies and the token formulas that
need to be satisfied for a rule match, even though Not is not relevant
to the latter.  But doing this is simpler than reimpementing functions
such as build_and.