Remove the deprecated 'have' call from bash-completion
The 'have' function has been deprecated in bash-completion, and its use
in the completion file is wrong. If 'spatch' is not available,
the completion function will not be declared, yet the 'complete' calls
relying on it will be run.

Instead, it is fine to assume that completion will be installed
alongside the program and can be enabled unconditionally as long as it
is installed.

Signed-off-by: Michał Górny <mgorny@gentoo.org>