pycocci: add requirements class
This will be used later to help examine different tools requirements.
For instance, this will be used to fine tune spatch operation depending
on the version of spatch you have. Later we'll also use this to add
support for SmPL patch equivalence proof support.

This introduces has no functional changes.

Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>