pycocci: figure out if to enable --jobs, --use-gitgrep or --use-glimpse
pycocci was originally written to set sensible defaults for doing
development with Coccinelle with a git tree and to also add
parallelism wrapper support. Coccinelle now has a '--jobs' argument
which does the same but it works better. This feature is available
as of 1.0.0 but will have some fixes which will go in on 1.0.2.

Likewise the 1.0.1 release has --use-gitgrep support which works
better than --use-coccigrep but has some issues on the 1.0.1 release
which are fixed on the 1.0.2 release. An old feature that has been
in place for a long time is --use-glimpse, and since we have that code
released now as open source we should provide support for it as
its expected we'd get the best results with it. The glimpse
code still needs a bit of work, as the public releases doesn't yet
build correctly, but nevertheless we should support it as glimpse
is now open sourced under the ISC license.

Instead of removing the old built-in wrapper for parallelism support
lets keep it and instead check the version of coccinelle you have
and let pycocci figure out what options you need.

For coccinelle 1.0.2, coccinelle will be able to let you default to
always use --use-gitgrep by default safely, it will call 'git grep'
for some cases, and if git was not installed or if your path was not
part of a git tree it will fallback to using --use-coccicheck. We can
optimize this a bit by checking both if you have git installed for you
and since we already have code which can easily check if your path is
part of a git tree we can check for that before using --use-gitgrep.
This will avoid unnecessary failed calls to git grep even though as
of coccinelle 1.0.2 it should be safe to default to it and use
coccinelle's fallback.

The check to see if a target path is part of a git tree is implemented
in gitname(), upon investingating git code setup_git_directory_gently()
is the shortest possible path to determine if a path is part of a git
tree, and respectively git rev-parse the easiest tool we can use to
infer this from a given path. Something similar is not implemented
but can be done for glimpse. This is not a priority however as glimpse
can't even build yet as-is with the new public release of code.

If all index options fail to check out we resort to --use-coccigrep as
the last-resort default.

Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>