Build system: make sure OCaml's version is at least 3.11.
This is required so that the position of tokens in C files can be
computed during the lexing stage, rather than during a dedicated stage
which involves reading the file once more.

More precisely, in order to update positions, the lexer makes use
of the Lexing.new_line function which has been introduced in OCaml 3.11.

This function does nothing magic, so all this could have been done
manually in the lexer, or one could have used conditionnally compiled
code to either do it ourselves on older version of OCamlor use the
function when available, but it just looked simpler to move on to
OCaml 3.11 which has been relesed more than five years ago now.