Python inizialization code: do not try to modify PYTHONPATH
Before initializing the Python interpreter, the code was modifying
the PYTHONPATH environment variable, adding ${HOME}/coccinelle to it.

However, it turns out that this modifiction was actually not taken into
account, because the Python interpreter has already been initialized
by the Pycaml interface.

This is a flaw in Pycaml that could be interesting to fix.