add WITH/OUT_TYPES annotation for exp and stmt metavar values
WITHOUT_TYPES allow successful matching of metavariables defined in ocaml
code, where type information is not available.  Note that if type
information is provided for ocaml generated metavariables, it must be
complete.  If some types are missing, the subexpressions will not match.
This is entirely the user's responsability; there is no check of this
property.