docs/manual/Makefile: improve the way hevea is called.
Hevea needs a file as an argument to the -exec option.
This file can be located in different directories.
The Makeile was looking for it in the recipe that produces the HTML and
this made Hevea'sinvocation quite complex.
It was thus hidden, which was not convenient at all.

This commit fixes this. Determining the file's path is now done in a
shell script whose result is put into a makevariable.

The commands used to invoke Hevea thus become simpler and can
hence be shown to the user.