Makefile: fix autotools pre ./configure version regression
The change to fold version information into a single helper
removed access to the VERSION variable before ./configure was
run. We have a slew of dependencies on this variable though,
users of this variable are typically only used for release
purpose so when the user hasn't yet run ./configure simply
assume we want the version information as if we're making
a release and don't want any dirty tree annotations.

Reported-by: Sébastien Hinderer <Sebastien.Hinderer@inria.fr>
Signed-off-by: Luis R. Rodriguez <mcgrof@suse.com>