When forall or when exists on ... is complicated, because the desired quantifier is not available to the context (typically make_seq).
The strategy is to have the ... return a formula containing the operator
XX, and then to clean up afterwards.  But the cleanup is only done from
EX when the argument of XX is an AU directly.  However, in the pattern
allowing a goto in a sequence and the rest of the pattern at the goto
label, a conjunction of AUs is generated, with the AU for the ... and
the AU for checking for gotos.  This pattern is extended to be able to
be wrapped in XX, and the cleanup function is extended to cleanup this
conjunction pattern.

For the cleanup, strict is chosen as the strict of the conjunction.
Strict has to do with how partial matches are constructed.  Since the
conjunction seems similar to the AX, it seems reasonable to use the same
strictness value.