parameterization for define parameter dot lists
Did a lot of work to allow #define parameters to be customized in the
Ast0 visitor, to allow unfavoring the commas, but not sure it's useful
because it's not currently possible to add things into the middle of a
on asingle line.  Perhaps this requirement will be removed someday and
the code will become useful.