Fixes issue #48.
This problem was due to the following issue in the OCaml compilers:

  http://caml.inria.fr/mantis/view.php?id=6661

This was actually already fixed by commits 33d2565da2cc, 4b27d7c6b404 and
b87e9ae2524190. This previous fix was however incomplete (working
only for the byte-code versoin of Coccinelle) and hackish.

The current fix also works for the native-code version of
Coccinelle and does not rely on Makefile and sed tricks.