Added: Compare standard and error outputs in tests.
In addition to the "<test>.res" file, tests may now have a
"<test>.out" file: if such a file exists, standard and error outputs
are compared to it. This may be useful to test the behavior of
scripts.

Note that if the "<test>.out" file exists, then the standard and error
outputs are redirected to the file "<test>.out.current" during the
execution of the test, to enable the comparison. This file is kept
afterwards.