pycocci: use os.path.abspath() for gitname
gitname() is currently used for SmPL <=> Patch
equivalence proof support. It uses git to test
for proofs and also show any differences found.
If the target directory passed where the user
wishes to do work on has is not part of a git
tree, it creates a git tree for you there.

When using relative paths the check for if a
target directory is part of a git tree currently
always fails, the consequences of which are that
if you try to show proof on an existing git tree
you'll end up with a sub-git project, and as such
a new .git under some directory under your target
path.

Correct this by using absolute paths.

Acked-by: Julia Lawall <julia.lawall@lip6.fr>
Signed-off-by: Luis R. Rodriguez <mcgrof@kernel.org>