Added: Python 3 support
New patch for Pycaml: python3-strings-classes.patch

pystring_as_string now converts to OCaml strings both unicode strings and
"bytes" strings.

NULL and the None value of Python are now represented by special pyobject
values that can be compared structurally in OCaml.

Pycaml.pyclass_init
  (classname: pyobject)
  (parents: pyobject)
  (fields: (string * pyobject list))
  (methods: (string * (pyobject -> pyobject) list))
  : pyobject
returns a new class type [classname] (a Python string) derived from the
classes [parents] (a tuple of Python strings), with the fields [fields]
and the methods [methods]. This function abstracts the differences
between the construction of a class type in Python 2 (with
[PyClass_new], where the dictionary can be populated after the call)
and Python 3 (with a call to [type()], where the dictionary should be
populated before the call).