Delete extra debug information.
For debugging strange comparison errors, it was useful to have
information about token types, and not just the strings, which could
have different token types depending on header file information.  But
this breaks the regression testing infrastructure, which requires the
exact error message to remain the same, so put the extra code under
comments.