Adjust categorization of free position variables.
Previously, we put the position any variables in the same category with
the negatied position variables.  This was not good for get_constants2,
because the any position metavariables induce a dependency while the
negated ones don't.  So a previous commit removed the any metavariables
from the negated list.  But the any ones and the negated ones do share a
property - in cocci.ml, we want to collect all of their values, and not
just the ones for the current environment instance.  Removing the any
metavariables from the negated position variables list meant that this
didn't happen, and the any annotation no longer had any effect.

Now free_vars keeps track separately of the negated position variables
and the any position metavariables.  Both are passed to get_constants2
and only the negated ones are use.  Both are unioned in cocci.ml.

Note that this change doesn't have to be propagated to prequel, because
prequel doesn't actually run semantic patches, which is the code in
cocci.ml that is affected by this change, it just uses the position
variable information in its version of get_constants2.