tests: add test to remove single statement from branch
At leaset with coccinelle 1.0.4 this test fails with
two issues, one is the return statement is shifted,
another is debatable -- an empty branch statement is
left. This is funtional however its debatable if this
is desirable by default.

v2: do not remove the if -- the if is needed as the items
    in the branch could consist of functional run time
    changes, for instance if (bar()).

Signed-off-by: Luis R. Rodriguez <mcgrof@kernel.org>