Python support for position constraints
A test case (and example) has been put in tests/python_poscon.cocci.
It is a straightforward translation of tests/poscon.cocci.

Python position constraints should be detected for Python
initialization. The code that did the detection for OCaml position
constraints was in Yes_prepare_ocamlcocci: the part of this code that
is common with Python has been moved to Parse_cocci (perhaps should we
find a better place for it?). Some uses of list concatenation have been
replaced with simple inline consing in Yes_prepare_ocamlcocci.

The constraint language set that is constructed in the module Cocci has
been switched to a StringSet (newly declared in the module Common) to
simplify the code (especially for unions).