Fixed: menhir and menhirLib always in sync
These changes ensure that the menhirLib used for linking is always in
sync with the menhir program that generates the parser.

SF Markus Elfring noticed that when menhir was installed in the system
but ocamlfind was not, the system menhir was used to compile the parser
but the bundled version of menhirLib was then used during linking, that
caused the build to fail when versions mismatched.

configure.ac now uses the menhir --suggest-menhirLib option to find the
menhirLib path if the option is available (as suggested by SF Markus
Elfring). Since this option is only available on recent versions of
menhir, configure.ac falls back to ocamlfind query menhirLib if
available, or else falls back to the bundled menhir both for compiling
and linking (in this case, the system menhir is just ignored).

Moreover, the Makefile rule that runs menhir to generate the .mli and
.ml from the .mly now depends on the Makefile.config file, to force the
parser regeneration if the configuration changes. (In addition, a
dependency of the .mli file to the .ml file has been added to prevent
the parallel compilation of both in case of multi-threaded compilation.)