Adjust position of drop_space_at_endline
It has to be called when the - code is still present, so that Coccinelle
can see that something has changed.  On the other hand it has to be
called after the pre-removal of - code, so that intermediate spaces and
newlines within the - code have been annotated as -.

Issue detected by Markus Elfring.