fix XX issue
XX is used to indicate a switch between forall and exists.  Cleanup does
something in reaction to this issue.  Cleanup expects EX to be composed
with an A operator, when there is an XX in the way, because only that
would represent a switch.  But at top level, if there are dots, and EX
is artifically added because there is a loop on enter or exit.  But this
EX is not taken into account in the addition of XX.  This move the
cleanup under the EX added by protect_top_level, to be in the expected
context.