Added: Support for --with-python and --python
The option --with-python is restored in configure to give a default
path/command for the Python interpreter.

The value can be overwritten with the --python option in the command
line.

Py.ml library has been updated to support custom Python interpreters. As
a side effect, the environment variable PYTHONHOME is now dynamically
set if it is not already defined. It should make the Python library work
even on systems where the environment variable has not been properly
set.