Fix default for COCCINELLE_HOME and PYTHONPATH
The default value for COCCINELLE_HOME was never correct.
The path of the executable (Sys.argv.(0)) is now used if this
directory contains standard.iso (typical case for development),
or else the path "../lib/coccinelle" relatively to the path of the
executable (typical case when Coccinelle is installed).

$COCCINELLE_HOME/python is added to PYTHONPATH if this path is not
already included.

These two changes make the spatch executable usable without
initialization script.