Adjust names used for script constraints.
Previously the names came from a global counter.  This was not safe if
the .cocci was reparsed due to iteration (necessary because of the
possibility to change the values of virtual identifiers).  New names
were generated, but at least in the case of ocaml, it seems that new
code was not generated, or at least not executed.  Now the names are
based on the rule name and metavariable name, which should make them
unique and unchanging.