Bitfield annotation on expression meta-variables
Answer to a request by Kumar Gala.
Expression meta-variables can be annotated with a bitfield (parsed as a
length: can be a constant or a meta-variable).
When an expression is annotated, it can only match with the projection
of a record field with the given bitfield alignment.
A FieldType constructor has been added to Ast_c.typeC to keep track of
bitfield alignment when typing projections.