correct production of new env
Fixes 4083bbcf

Apply_cocci_rule starts with an incoming environment, extends it for the
negated position variables and the position any variables, and matches
the rule.  Afterwards, it has to create a new env with the still
relevant bindings.  This was done by taking all of the still relevant
variables that were not in the original environment.  But some elements
of that environment could be complete extended, ie not existing at all
in the original environment.  We want to drop those bindings completely.
Indeed, the only new variables kept should be the ones that are local to
the current rule.

The expected output for the new test case is:

k: lock22.c:5
v: lock22.c:6
k: lock22.c:7
v: lock22.c:6