better parsing results
Sometimes a parsing error is detected inside parentheses.  In this case,
when there is an identifier just before the parentheses, we can
dynamically convert that identifier name to a macro, that just pases
over the arguments.  This allows the enclosing function (or other
toplevel entity) to be parsed.

The main changes are in parse_c, where the line and column of the error
are recorded, and in cpp_token_c where the function around detects that
some parentheses surround the error point.