Be more restrictive about declaring a word to be a single line macro
A word is likely a single line macro only if the previous line ended
with a statement end marker, namely ; { } or :.  The goal is to avoid
cases where random words are considereed to be macrodecls just because
they appear alone on a line due to indentation issues.

An example could be PH_ACTIVATE in drivers/isdn/hisax/hfc_usb.c.