enable caching of header files
Hunk 1: When we retrieve a cached header file, we get both the file and
the list of things that it includes.  We need to include all of them
too.

Hunk 2: Before caching a parsed header file, collect the other header
files that it includes.

Hunks 3 and 4: Make handle_include a freestanding function so that it
can be used when a header file is retrieved from the cache.

Hunk 5: Rewrite the existing call to handle_include to call the new
freestanding function.