more control over K&R parsing
A prototype such as static void authenticate_timeout(u_long); can be
interpreted as having a single parameter named u_long, rather than
having a single parameter of type u_long.

The default is now that if a non-KR parameter has been seen in the
current file, then all subsequent parameters are considered to be
non-KR, and u_long will be classified as a typedef.

--force-kr prevents this change of state; the possibility that a
parameter is a K&R parameter is always considered.

--prevent-kr causes all parameters to be considered non-K&R.  It is
needed when a typedef is never used with a variable name before its
first use in a parameter list.