step 2 towards local variables in constraint scripts
Allow arbitrary code in constraint scripts.

Add information to all_metavariasbles as metavariables are declared in
order to have access to the metadecs in any subsequent constraint
scripts.  The metadecls are neede to know the metavariable type.

Script parameters are no longer restricted to checked metanames (ie
names with a rule name), but now allow accesses to previously defined
metavariables.