remove possibility of double space
When the only modification is to add tokens after all of the minus
tokens, space at the end of the minus tokens will come out as being at
the front of the added tokens, where it is probably not wanted.  So
allow that space to be dropped.  Only preserve trailing space if the
added tokens are before or within the dropped ones.  this is detected by
exists_before_end.  See the test pcim.{c,cocci}.