remove conflict between local python script variables and virtuals
Iteration is implemented using constructs like:

it.add_virtual_rule(after_start)
it.add_virtual_identifier(y, vl)

To support this, after_start and y have to be bound to something.  In
OCaml, we use constructors in these positions, but they are not
available in python.  An approach would be to require strings around
these names, but that would be ugly.  The strategy taken is to by
default bind the names in a python script to the string value of the
name.  This is done after the local bindings for the script.  But this
causes a problem if the script has a local variable with the same name
as a virtual rule or identifier, because the proper value is
overwritten.  Thus, only create the binding if there is no local
variable (ie no variable imported from another rule) with the same name.