change neighbors of if etc
If, for and while have a fixed number of expected neighbors that are all
disjoint.  Matching them with and should cause abort in the case of
failure more quickly and thus be more efficient.  For switch, on the
other hand, we are stuck with a disjuntion over all of the neighbors
because we don't know how many branches there are.  This is expected to
improve the performance of the rule inif in the kernel semantic patch
mini_lock.cocci.